requirements.txt
Django==4.0.4
mysqlclient==2.1.1
Pillow==10.0.0
pdfkit==1.0.0
python-dotenv==1.0.0
django-crispy-forms==2.0
sslcommerz-python==0.0.7
python-dateutil==2.8.2
requests==2.31.0
django-cleanup==8.0.0
django-widget-tweaks==1.5.0